app: provide in-tree board overlay
Provide an overlay to shows how to run the example-application on boards
that are part of the Zephyr tree. Sometimes initial development is
performed on dev-kits, so it can be useful. A different vendor than the
one used for the custom board (Nordic) has been chosen to show
application portability.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>