example-app: twister-verifiable samples/tests of subsys module_hello
This change provides CI-discoverable and verifiable content in the
`include/`, `subsys/`, `samples/`, and `tests/` directories of the
Zephyr tree.  `testcase.yaml` and `sample.yaml` files are discoverable
and runnable by `twister -T example-application`. The samples and
tests themselves verify the `include/` tree is visble through the
build system and that the subsystem is buildable and runnable.
`module.yml` has been configured so the `testcase.yaml` and
`sample.yaml` should be automatically discoverable without explicit
`twister` parameters.
.
This approach was chosen to provide examples in Zephyr directory
sub-trees not already provided, to provide twister-runnable tests
and samples (not already provided), and to provide valid API and
subsys documentation files as a test case for Zephyr's document
generation tools.
.
It appears no assumptions were made in this design.
.
The implementation was verified in the following ways:
  - Ran
        `./zephyr/scripts/twister -v \
         --tag extensibility --tag introduction \
         -T example-application/ \
         --dry-run`
    - Verified that `twister-out/twister.log` found possible
      test cases under `example-application/tests/` and
      `example-application/samples/`, and did not discard
      all of them due to filters.
  - Ran the above command without `--dry-run`.
    - Verified that it reported PASSED on >=1 build-only `samples/` job
      (PASSED 5/5).
    - Verified that it reported PASSED or FAILED Timeout on
      >=1 non-build-only `samples/` job  (FAILED Timeout on 13/16).
    - Verified that it reported PASSED on >=1 non-build-only `tests/` job.
      (It passed 18/21, including running on qemu_cortex_m0.)
  - Ran
        `   cp -r example-application/include/module_hello.h \
                  zephyr/include/ \
         && cp -r example-application/samples/subsys/module_hello \
                  zephyr/samples/subsys/ \
         && cd zephyr/doc && make clean && make && cd ../.. \
         && rm -f zephyr/include/module_hello.h \
         && rm -rf zephyr/samples/subsys/module_hello \
        `
    - Verified "module_hello_said" was found in the generated documentation,
      confirming the header was valid if it were in the Zephyr repo.
    - Verified the "Module Hello" sample was found in the generated document,
      confirming the README.txt is valid.

Signed-off-by: Gregory Shue <gregory.shue@legrand.us>