modules: Set module name
Follow the Zephyr recommendation to explicitly set the module name in
`module.yml`. Add a build check in the module-level CMakeLists.txt to
verify the value of the build variable ${module_name} is the expected
value so that other modules may reference this one without needing to
know the path to this module.

Verified by:
1. Successfully building
   `west build -p always -b custom_plank example-application/app`
   with the name check and the defaulted name.
2. Unsuccessfully buildng
   `west build -p always -b custom_plank example-application/app`
   with the name check and an unexpected name set in `module.yml`
3. Unsuccessfully buildng
   `west build -p always -b custom_plank example-application/app`
   with the name check and the expected name set in `module.yml`

Fix #50866

Signed-off-by: Gregory SHUE <gregory.shue@legrand.com>