doc: remove breathe
Breathe is no longer used in Zephyr, mainly to its poor performance and
lack of proper maintainership. Therefore, do not recommend it in our
template application. Previous external C references do not work either,
so let's remove them. We may polish doxybridge/runner in the future and
make it a published extension on PyPI so that 3rd party repos like
applications can use it.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>